SourceCodester Yoga Class Registration System
SourceCodester Yoga Class Registration System 1.0- SQLi 
A vulnerability was found in SourceCodester Yoga Class Registration System 1.0. It has been declared as critical. Affected by this vulnerability is an unknown functionality of the component Delete User. The manipulation of the argument id leads to sql injection. This vulnerability is known as CVE-2023-0981. The attack can be launched remotely. Furthermore, there is an exploit available.